+++ src/index.ts
@@ -6,7 +6,7 @@
 export * from './md-elements.js';
 export * from './md-icons.js';
 export * from './regexps.js';
-export * from './imports.js';
+export * from './stylesheet/stylesheet.js';
 export * from './fonts.js';
 
 export * from './md-elements-imports-map.js';